Added long fix

+++ utils/train.py
@@ -29,8 +29,7 @@ class Trainer:
         return ((input_sequences == output_sequences).sum(axis=1) / float(self.FIXED_PROTEIN_LENGTH)).mean()
 
     def __inner_iteration(self, x, training: bool):
-        x = x.to(self.device)
-
+        x = x.long().to(self.device)
 
         # update the gradients to zero
         if training:
@@ -41,7 +40,7 @@ class Trainer:
         # predicted = predicted.view(1,predicted.shape[0], -1)
         # reconstruction loss
         # predicted = F.log_softmax(predicted, 1)
-        recon_loss = self.criterion(predicted, x.long())
+        recon_loss = self.criterion(predicted, x)
 
         loss = recon_loss.item()
         # reconstruction accuracy
